+++ backend/app/services/providers/google/health_api/sleep.py
@@ -24,17 +24,16 @@ from app.services.event_record_service import event_record_service
 from app.services.providers.api_client import make_authenticated_request
 from app.services.providers.google.health_api.helpers import (
     GOOGLE_HEALTH_API_SOURCE,
-    as_int,
     extract_source,
     parse_interval,
     parse_rfc3339,
     physical_interval,
     read_number,
-    to_decimal,
     zone_offset_from,
 )
 from app.services.providers.templates.base_oauth import BaseOAuthTemplate
 from app.services.raw_payload_storage import store_raw_payload
+from app.utils.conversion import as_int, to_decimal
 
 
 class GoogleHealthApiSleep:
