+++ backend/app/services/providers/strava/workouts.py
@@ -23,6 +23,7 @@ from app.services.event_record_service import event_record_service
 from app.services.providers.strava.coverage import STREAM_KEY_SERIES_TYPE, STREAM_KEYS_PARAM
 from app.services.providers.templates.base_workouts import BaseWorkoutsTemplate
 from app.services.timeseries_service import timeseries_service
+from app.utils.conversion import kilojoules_to_kcal
 from app.utils.dates import offset_to_iso
 from app.utils.sentry_helpers import log_and_capture_error
 from app.utils.structured_logging import log_structured
@@ -183,7 +184,7 @@ class StravaWorkouts(BaseWorkoutsTemplate):
         if raw_workout.calories is not None and raw_workout.calories > 0:
             metrics["energy_burned"] = Decimal(raw_workout.calories)
         elif raw_workout.kilojoules is not None:
-            metrics["energy_burned"] = Decimal(raw_workout.kilojoules) * Decimal("0.239")  # convert to kcal
+            metrics["energy_burned"] = kilojoules_to_kcal(raw_workout.kilojoules)
 
         # Moving time
         if raw_workout.moving_time is not None:
